+++ src/cli/command-registrar-builder.js
@@ -27,6 +27,7 @@ import Fetch from './commands/private-cmds/_fetch-cmd';
 import Log from './commands/public-cmds/log-cmd';
 import Build from './commands/public-cmds/build-cmd';
 import EnvsAttach from './commands/public-cmds/envs-attach-cmd';
+import EjectConf from './commands/public-cmds/eject-conf-cmd';
 import CiUpdate from './commands/private-cmds/ci-update-cmd';
 import RefreshScope from './commands/private-cmds/refresh-scope-cmd';
 import CatScope from './commands/private-cmds/cat-scope-cmd';
@@ -78,6 +79,7 @@ export default function registerCommands(extensionsCommands): CommandRegistrar {
       new Fetch(),
       new Build(),
       new EnvsAttach(),
+      new EjectConf(),
       new DescribeScope(),
       new CiUpdate(),
       new RefreshScope(),