+++ src/cli/templates/help.js
@@ -47,6 +47,7 @@ const baseTemplate = (extensionsTemplate) => {
     'test'
   )}       test any set of components with configured tester (component tester or as defined in bit.json)
   ${chalk.cyan('  envs-attach')}attach workspace environments to components
+  ${chalk.cyan('  eject-conf')} ejecting components configuration
 
   ${chalk.underline('general commands')}
     ${chalk.cyan('login')}      log the CLI into bitsrc.io