+++ src/consumer/bit-map/exceptions/invalid-config-dir.js
@@ -1,0 +1,11 @@
+/** @flow */
+import AbstractError from '../../../error/abstract-error';
+
+export default class InvalidConfigDir extends AbstractError {
+  compId: string;
+
+  constructor(compId: string) {
+    super();
+    this.compId = compId;
+  }
+}