+++ src/extensions/tester-extension.js
@@ -29,15 +29,10 @@ export default class TesterExtension extends EnvExtension {
    * used for ejecting env for imported component
    * @param {*} param0
    */
-  async writeFilesToFs({
-    bitDir,
-    ejectedEnvsDirectory
-  }: {
-    bitDir: string,
-    ejectedEnvsDirectory: string
-  }): Promise<string> {
+  async writeFilesToFs(baseOpts: { configDir: string, deleteOldFiles: boolean }): Promise<string> {
     Analytics.addBreadCrumb('tester-extension', 'writeFilesToFs');
-    return super.writeFilesToFs({ bitDir, ejectedEnvsDirectory, envType: this.envType });
+    const concreteOpts = { ...baseOpts, envType: this.envType };
+    return super.writeFilesToFs(concreteOpts);
   }
 
   /**